refactor chatroom component
