+++ client/src/components/views/Chatroom/Chatroom.css
@@ -42,16 +42,20 @@
 }
 
 .buttons {
-    cursor: default;
+    cursor: pointer;
     border: none;
+    outline: none;
     border-radius: 50px;
     width: 178px;
     height: 57px;
     color: white;
-    background: #629ff4;
+    background: #708fc1;
     text-transform: uppercase;
     font-family: Open Sans, sans-serif;
     font-weight: 600;
+}
+.buttons:hover{
+    background: #629ff4;
 }
 .pill{
     border-radius: 29px;
@@ -99,6 +103,7 @@ button{
     background: none;
 
     cursor: pointer;
+    outline: none;
 }
 button.play:hover path{
     fill: #59cbb7;
@@ -121,68 +126,3 @@ button.share:hover path{
 
     color: #4a4a4a;
 }
-.pill.done{
-    /*background: white;*/
-    /*border: solid 2px grey;*/
-}
-.pill.contents{
-    /*display: flex;*/
-    /*height: 100%;*/
-    /*align-items: center;*/
-}
-/*.pill.recording.done.contents>{*/
-/*    height: 100%;*/
-/*}*/
-/*.pill.recording.placeholder{*/
-/*    width: 110px;*/
-/*}*/
-/*.pill {*/
-/*    border-radius: 29px;*/
-/*    box-sizing: border-box;*/
-
-/*    !*width: 46px;*!*/
-/*    !*height: 46px;*!*/
-
-/*    !*flex-shrink: 0;*!*/
-/*    !*display: flex;*!*/
-/*    !*flex-direction: row-reverse;*!*/
-/*    !*justify-content: center;*!*/
-/*    !*align-items: center;*!*/
-
-/*    !*overflow: hidden;*!*/
-/*}*/
-
-/*.pill .done {*/
-/*    background: white;*/
-/*    border: none;*/
-/*    padding-left: 6px;*/
-/*}*/
-
-/*.pill .done .contents {*/
-/*    display: flex;*/
-/*    height: 100%;*/
-/*    align-items: center;*/
-/*}*/
-
-/*.play {*/
-/*    border: none;*/
-/*    padding: 0;*/
-/*    height: 100%;*/
-/*}*/
-
-/*.padder {*/
-/*    display: inline-block;*/
-/*    padding: 0 16px;*/
-/*    border-left: none;*/
-/*}*/
-
-/*.num {*/
-/*    border: 1px solid #f5f4f3;*/
-/*    margin: 0;*/
-/*    background: #f3f2f1;*/
-/*}*/
-
-/*.button {*/
-/*    background: none;*/
-/*    cursor: pointer;*/
-/*}*/
