+++ client/src/components/views/Chatroom/Section/RecordButton.js
@@ -1,29 +1,26 @@
-import React, {useState} from 'react';
+import React from 'react';
 import './RecordButton.css';
 import ReactRecord from 'react-record';
 import {StopIcon,MicIcon} from '../../../ui/icons';
 
-export default function RecordButton({sendDataFromChild}) {
-
-    const [isRecording, setIsRecording] = useState(false)
-
+export default function RecordButton(props) {
     return (
         <div style={{margin: '4rem auto'}}>
             <div className="primary-button">
                 <ReactRecord
-                    record={isRecording}
+                    record={props.isRecording}
                     onData={() => {}}
                     onSave={() => {}}
                     onStart={() => {}}
                     onStop={blobObject => {
-                        // console.log('blobObject is: ', blobObject);
-                        sendDataFromChild(blobObject);
+                        console.log('blobObject is: ', blobObject);
+                        props.setAudio(blobObject);
                     }}>
-                    {isRecording ?
-                        <button onClick={() => setIsRecording(false)} className="primary-button button" type="button">
+                    {props.isRecording ?
+                        <button onClick={() => props.setIsRecording(false)} className="primary-button button" type="button">
                             <StopIcon/>
                         </button> :
-                        <button onClick={() => setIsRecording(true)} className="primary-button button" type="button">
+                        <button onClick={() => props.setIsRecording(true)} className="primary-button button" type="button">
                             <MicIcon/>
                         </button>}
                 </ReactRecord>
