+++ client/src/components/views/Chatroom/Section/Wave.js
@@ -1,4 +1,3 @@
-import {useRef,useEffect} from 'react';
 // wave
 const RATIO = window.devicePixelRatio;
 const WIDTH = RATIO * 320;
